Capstone_Project/app/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
feedparser==5.2.1
Flask==1.1.2
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
nbapy==1.1.8
numpy==1.19.1
pandas==1.1.0
plotly==4.9.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-cache==0.5.2
retrying==1.3.3
rq==1.5.1
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
